+++ blog-common/src/main/java/com/fleet/common/service/impl/BaseServiceImpl.java
@@ -25,7 +25,7 @@ public abstract class BaseServiceImpl<T> implements BaseService<T> {
     }
 
     @Override
-    public Boolean deletes(Integer[] ids) {
+    public Boolean deletes(Long[] ids) {
         return baseDao().deletes(ids) != 0;
     }
