+++ blog-common/src/main/java/com/fleet/common/service/menu/MenuService.java
@@ -10,7 +10,7 @@ public interface MenuService extends BaseService<Menu> {
     /**
      * 获取所有 menuId 集合（包括）
      */
-    List<Integer> idList(Integer id);
+    List<Long> idList(Long id);
 
     List<Menu> buildTree(List<Menu> menuList);
 }
