+++ blog-common/src/main/java/com/fleet/common/service/role/RoleMenuService.java
@@ -8,9 +8,9 @@ import java.util.List;
 
 public interface RoleMenuService extends BaseService<RoleMenu> {
 
-    List<Integer> menuIdList(Integer roleId);
+    List<Long> menuIdList(Long roleId);
 
-    List<Menu> menuList(Integer roleId);
+    List<Menu> menuList(Long roleId);
 
-    List<String> permitList(Integer roleId);
+    List<String> permitList(Long roleId);
 }
