+++ blog-common/src/main/java/com/fleet/common/service/role/RoleService.java
@@ -10,7 +10,7 @@ public interface RoleService extends BaseService<Role> {
     /**
      * 获取所有 roleId 集合（包括）
      */
-    List<Integer> idList(Integer id);
+    List<Long> idList(Long id);
 
     List<Role> buildTree(List<Role> roleList);
 }
