+++ blog-provider/provider-admin/src/main/java/com/fleet/provider/admin/dao/DeptDao.java
@@ -14,5 +14,5 @@ import java.util.List;
 @Repository
 public interface DeptDao extends BaseDao<Dept> {
 
-    List<Integer> idList(Dept dept);
+    List<Long> idList(Dept dept);
 }
